Update about.md
Thought it would be nice to use the name 'tus' in the title. I am the type that usually had 10+ tabs open in his browser and I doubt I am alone in that. So I think it's good to have the name there, so people constantly see it from the corner of their eye!