update subscribe url (#203)
* update subscribe url

https://app.intercom.com/a/apps/qiqpfgjg/inbox/inbox/5079389/conversations/26852700005173, no.1i
On line 199, the link, "https://transloadit.us1.list-manage.com/subscribe/post?u=98e560c614f2baaf47237f500&amp;id=37aafbe5c1" does not work when it is pressed. It redirects to a 404:Page Not Found. Is there another URL that we can replace it with? If so, I will make the changes for you

* removed subscription block

* removed whole subscription section

forgot to remove the whole section.